moved new `'android_locale'` tests into tests/bin
